separated between ugbio_core and ugvc
