+++ test/resources/unit/joint/test_gvcf_bed/calls.expected.bed
