+++ test/resources/unit/joint/test_gvcf_bed/calls.g.vcf.gz
